requirements.txt
httpx>=0.27
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
anthropic>=0.40
openpyxl>=3.1
python-dotenv>=1.0
streamlit>=1.30
plotly>=5.18
